+++ src/ezdxf/addons/gerber_D6673.py
@@ -1,10 +1,6 @@
 # Copyright (c) 2022, Manfred Moitzi
 # License: MIT License
 # This add-on was created to solve this problem: https://github.com/mozman/ezdxf/discussions/789
-"""
-This add-on creates special DXF files for use by Gerber Technology applications which
-have a broken DXF loader.
-"""
 from __future__ import annotations
 from typing import TextIO
 import os
@@ -18,8 +14,8 @@ __all__ = ["export_file", "export_stream"]
 
 
 def export_file(doc: Drawing, filename: str | os.PathLike) -> None:
-    """Exports the specified DXF R12 document, intended to contain content conforming to
-    the ASTM-D6673-10 standard, in a special way so that Gerber Technology applications
+    """Exports the specified DXF R12 document, which should contain content conforming
+    to the ASTM-D6673-10 standard, in a special way so that Gerber Technology applications
     can parse it by their low-quality DXF parser.
     """
     fp = io.open(filename, mode="wt", encoding="ascii", errors="dxfreplace")
@@ -27,6 +23,8 @@ def export_file(doc: Drawing, filename: str | os.PathLike) -> None:
 
 
 def export_stream(doc: Drawing, stream: TextIO) -> None:
+    """Exports the specified DXF R12 document into a `stream` object."""
+
     if doc.dxfversion != ezdxf.const.DXF12:
         raise ezdxf.DXFVersionError("only DXF R12 format is supported")
     tagwriter = TagWriter(stream, write_handles=False, dxfversion=ezdxf.const.DXF12)
@@ -50,11 +48,11 @@ def _export_blocks(doc: Drawing, tagwriter: AbstractTagWriter) -> None:
     # This is the important part:
     #
     # Gerber Technology applications have a low-quality parser which do not accept DXF
-    # files that contain blocks without ASTM-D6673-10 content, such as the *MODEL_SPACE
-    # and *PAPER_SPACE blocks.
+    # files that contain blocks without ASTM-D6673-10 content, such as the $MODEL_SPACE
+    # and $PAPER_SPACE blocks.
     #
     # This is annoying but the presence of these blocks is NOT mandatory for
-    # the DXF Standard.
+    # the DXF R12 standard.
     #
     tagwriter.write_str("  0\nSECTION\n  2\nBLOCKS\n")
     for block_record in doc.block_records: